Split content from wfdbio.cc and wfdblib.h into new netfiles.cc and netfiles.hh
